Move arbitrary module to testutils
