+++ soroban-sdk/src/lib.rs
@@ -786,7 +786,6 @@ pub mod xdr;
 
 pub mod testutils;
 
-pub mod arbitrary;
 mod arbitrary_extra;
 
 mod tests;
